Modify when to start countdown to MAP email
An order can be created weeks before it reaches the status configured in
magento backend. A state can occur where an MAP email is sent
immediately after the order gets the desired status, which is likely
before the customer has received their order. If the desired status is
"complete" or "shipped", it's likely because we want to withhold these
emails until the goods are delivered. Using the creation date of the
shipment instead of the creation date of the order gets us closer to
this behavior.