fixed the review export not working -
a file is created, but then the data is regenerated and used to send to the user. However this is broken as the $data
variable in exportData only ever contains the last line of data. There is no need to generate it twice.

this change removes the exportData function and just reads the contents of the already created csv. The export now works properly.